refactor: move functionality from dash_app to metrics
