+++ app/metrics/templates/metrics/model-form.html
@@ -6,7 +6,7 @@
 {% block content %}
     <h1>{{title}}</h1>
     {% include 'common/tabs.html' %}
-    {% include 'dash_app/data-warning.html' %}
+    {% include 'metrics/data-warning.html' %}
     <h2>Stats</h2>
     <div class="mb-3 row">
         <table class="table table-bordered">
