+++ app/metrics/templates/metrics/model-list.html
@@ -6,7 +6,7 @@
 {% block content %}
     <h1>{{title}}</h1>
     {% include 'common/tabs.html' %}
-    {% include 'dash_app/data-warning.html' %}
+    {% include 'metrics/data-warning.html' %}
     <div class="mb-3 row">
         <nav aria-label="Page navigation">
             <ul class="pagination">
