+++ app/metrics/templates/metrics/upload.html
@@ -6,7 +6,7 @@
 {% block content %}
     <h1>{{ title }}</h1>
     {% include 'common/tabs.html' %}
-    {% include 'dash_app/data-warning.html' %}
+    {% include 'metrics/data-warning.html' %}
     {% for form in forms %}
     <h2>{{form.title}}</h2>
     {% if form.description %}{% for paragraph in form.description %}<p>{{ paragraph }}</p>{%endfor%}{% endif %}
