+++ app/metrics/urls.py
@@ -3,9 +3,39 @@ from django.urls import path
 
 from metrics import views
 from metrics.forms import UserLoginForm
+from metrics.views.upload import upload_data
+from metrics.views.model_views import (
+    EventView,
+    InstitutionView,
+    EventListView,
+    InstitutionListView,
+    QualityMetricsDeleteView,
+    DemographicMetricsDeleteView,
+    ImpactMetricsDeleteView,
+)
 
 urlpatterns = [
-    path('login/', LoginView.as_view(template_name='metrics/login.html',
-                                     authentication_form=UserLoginForm, next_page='/'), name='login'),
+    path('login/', LoginView.as_view(
+        template_name='metrics/login.html',
+        authentication_form=UserLoginForm, next_page='/'),
+        name='login'
+    ),
     path('logout/', LogoutView.as_view(next_page='/'), name='logout'),
+    path('upload-data', upload_data, name='upload-data'),
+    path('event/<int:pk>', EventView.as_view(), name='event-edit'),
+    path('institution/<int:pk>', InstitutionView.as_view(), name='institution-edit'),
+    path('event/list', EventListView.as_view(), name='event-list'),
+    path('institution/list', InstitutionListView.as_view(), name='institution-list'),
+    path('event/delete-metrics/demographic/<int:pk>',
+        DemographicMetricsDeleteView.as_view(),
+        name="demographic-delete-metrics"
+    ),
+    path('event/delete-metrics/impact/<int:pk>',
+        ImpactMetricsDeleteView.as_view(),
+        name="impact-delete-metrics"
+    ),
+    path('event/delete-metrics/quality/<int:pk>',
+        QualityMetricsDeleteView.as_view(),
+        name="quality-delete-metrics"
+    )
 ]
