+++ app/metrics/views/common.py
